EditTimeの引数の取り方について確認
