+++ src/components/Home/Daily.tsx
@@ -3,28 +3,20 @@ import {
   View,
   Text,
   RefreshControl,
-  Button,
-  TextInput,
+  TouchableOpacity,
 } from "react-native";
 import { useState, useEffect } from "react";
 import { useTailwind } from "tailwind-rn/dist";
 import { useDispatch, useSelector } from "react-redux";
 import { AppDispatch } from "@stores/index";
-import { RootReducer } from "../../App";
-import {
-  getMonthlyHistories,
-  History,
-  deleteHistory,
-  changeMeansuringTime,
-  updatePrimary,
-} from "@stores/history";
+import { RootReducer } from "../../../App";
+import { getDailyHistories, History } from "@stores/history";
 import { VictoryPie } from "victory-native";
 import DropDownPicker from "react-native-dropdown-picker";
 import { dateFormat } from "@utils/format";
-import ChangeInfo from "./ChangeInfo";
-import Modal from "react-native-modal";
-import { Picker } from "@react-native-picker/picker";
-import { PrimaryCategory, getPrimaries } from "@stores/categories";
+import { getPrimaries } from "@stores/categories";
+import { useNavigation, NavigationProp } from "@react-navigation/native";
+import { RootStackParamList } from "../../../App";
 
 type item = {
   label: string;
@@ -43,12 +35,14 @@ export default () => {
   const {
     user: { user },
     history: {
-      histories: { monthly },
+      histories: { daily },
     },
-    categories: { primaryCategories },
   } = useSelector((store: RootReducer) => store);
 
-  const monthlyMap = monthly.reduce(
+  const navitaoin =
+    useNavigation<NavigationProp<RootStackParamList, "HomeTop">>();
+
+  const dailyMap = daily.reduce(
     (
       prev: {
         [key: string]: {
@@ -77,12 +71,7 @@ export default () => {
     },
     {}
   );
-
-  const [modalType, setModalType] = useState<
-    "initial" | "editTime" | "updatePrimary" | null
-  >(null);
-  const [selectPrimary, setSelectPrimary] = useState();
-  const [pieData, setPieData] = useState<pie[]>(Object.values(monthlyMap));
+  const [pieData, setPieData] = useState<pie[]>(Object.values(dailyMap));
   const [open, setOpen] = useState(false);
   const [primary, setPrimary] = useState(null);
   const [primaries, setPrimaries] = useState<item[]>([]);
@@ -94,96 +83,20 @@ export default () => {
     dispatch(getPrimaries({ userID: user!.id }));
   }, [user]);
 
-  const close = () => setModalType(null);
-
-  const InitialModal = ({
-    close,
-    editTime,
-    deleteItem,
-    showPrimary,
-  }: {
-    close: () => void;
-    editTime: () => void;
-    deleteItem: () => void;
-    showPrimary: () => void;
-  }) => {
-    const tailwind = useTailwind();
-    return (
-      <>
-        <View style={tailwind("bg-white p-2 m-1 rounded-2xl")}>
-          <Text style={tailwind("text-center text-base")}>
-            データの修正はこちらから
-          </Text>
-          <Button title="カテゴリ情報の修正" onPress={showPrimary} />
-          <Button title="計測時間の修正" onPress={editTime} />
-          <Button
-            title="データの削除"
-            onPress={() => {
-              close();
-              deleteItem();
-            }}
-          />
-        </View>
-      </>
-    );
-  };
-  const ShowPrimaryPicker = ({ id }: { id: string }) => {
-    return (
-      <Picker
-        selectedValue={selectPrimary}
-        onValueChange={(item) => setSelectPrimary(item)}
-      >
-        {primaryCategories.map((data) => (
-          <Picker.Item label={data.name} value={data.name} />
-        ))}
-      </Picker>
-    );
-  };
-
-  const EditTimeModal = ({ id, close }: { id: string; close: () => void }) => {
-    const [time, setTime] = useState("");
-    const tailwind = useTailwind();
-    return (
-      <View style={tailwind("bg-white p-12 m-1 rounded-2xl")}>
-        <View>
-          <Text style={tailwind("font-bold text-center")}>計測時間を修正</Text>
-        </View>
-        <TextInput
-          style={tailwind("border rounded-md px-2 py-1 mt-1")}
-          value={time}
-          onChangeText={(text) => setTime(text)}
-          autoFocus
-        />
-        <Button
-          title="OK"
-          onPress={() => {
-            close;
-            dispatch(
-              changeMeansuringTime({
-                historyId: id,
-                measuringTime: time,
-              })
-            );
-          }}
-        />
-      </View>
-    );
-  };
-
   useEffect(() => {
-    if (user) dispatch(getMonthlyHistories({ userId: user!.id }));
+    if (user) dispatch(getDailyHistories({ userId: user!.id }));
   }, [user]);
   useEffect(() => {
-    const primaryInfo = Object.values(monthlyMap).map((item) => {
+    const primaryInfo = Object.values(dailyMap).map((item) => {
       return { label: item.x, value: item.id };
     });
     setPrimaries([...primaryInfo, { label: "全て", value: "all" }]);
-    setPieData(Object.values(monthlyMap));
-  }, [monthly]);
+    setPieData(Object.values(dailyMap));
+  }, [daily]);
 
   useEffect(() => {
     if (primary && primary !== "all") {
-      const secondaryMap = monthlyMap[primary].secondaries.reduce(
+      const secondaryMap = dailyMap[primary].secondaries.reduce(
         (
           pre: {
             [key: string]: {
@@ -225,8 +138,8 @@ export default () => {
     }
 
     if (primary === "all") {
-      setPieData(Object.values(monthlyMap));
-      const newSecondaries = monthly
+      setPieData(Object.values(dailyMap));
+      const newSecondaries = daily
         .filter(
           (x, i, array) =>
             array.findIndex((y) => y.secondary_id === x.secondary_id) === i
@@ -244,7 +157,10 @@ export default () => {
   const renderItem = ({ item }: { item: History }) => {
     const timeinfo = Number(item.measuring_time) / 60;
     return (
-      <View style={tailwind("flex items-center")}>
+      <TouchableOpacity
+        style={tailwind("flex items-center")}
+        onPress={() => navitaoin.navigate("HistoryDetail", { item: item })}
+      >
         <View style={tailwind("ml-2 pl-1 w-4/5")}>
           <Text style={tailwind("text-base")}>
             {dateFormat(item.created_at.toDate())}
@@ -259,31 +175,12 @@ export default () => {
             <Text style={tailwind("text-base font-bold")}>
               {item.primary_name}
             </Text>
-            <ChangeInfo onPress={() => setModalType("initial")} />
-            <Modal isVisible={Boolean(modalType)} onBackdropPress={close}>
-              {modalType === "initial" && (
-                <InitialModal
-                  close={close}
-                  editTime={() => setModalType("editTime")}
-                  deleteItem={() =>
-                    dispatch(deleteHistory({ historyId: item.id }))
-                  }
-                  showPrimary={() => setModalType("updatePrimary")}
-                />
-              )}
-              {modalType == "editTime" && (
-                <EditTimeModal id={item.id} close={close} />
-              )}
-              {modalType == "updatePrimary" && (
-                <ShowPrimaryPicker id={item.id} />
-              )}
-            </Modal>
           </View>
           <Text style={tailwind("text-base font-bold text-right mr-1 pr-1")}>
             {timeinfo.toFixed(2)}min
           </Text>
         </View>
-      </View>
+      </TouchableOpacity>
     );
   };
 
@@ -327,15 +224,13 @@ export default () => {
         />
       </View>
       <FlatList
-        data={monthly}
+        data={daily}
         renderItem={renderItem}
         keyExtractor={(item) => item.id}
         refreshControl={
           <RefreshControl
             refreshing={false}
-            onRefresh={() =>
-              dispatch(getMonthlyHistories({ userId: user!.id }))
-            }
+            onRefresh={() => dispatch(getDailyHistories({ userId: user!.id }))}
           />
         }
       />
